# perf(es/minifier): Make graph-analysis of DCE single-pass