--- a/crates/swc_ecma_transforms_optimization/src/simplify/dce/mod.rs
+++ b/crates/swc_ecma_transforms_optimization/src/simplify/dce/mod.rs
@@ -1,3 +1,5 @@
+mod vars;
+
 use std::{borrow::Cow, sync::Arc};
 
 use indexmap::IndexSet;