… crates/swc_ecma_transforms_optimization/src/simplify/dce/vars.rs